ctxcat: Rename aictx -> ctxcat
