+++ ctxcat/main.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 	"os"
 
-	aictx "github.com/Unique-Divine/jiyuu/aictx/src"
+	ctxcat "github.com/Unique-Divine/jiyuu/ctxcat/src"
 )
 
 func main() {
-	appCmd := aictx.NewAppCmd()
+	appCmd := ctxcat.NewAppCmd()
 	if err := appCmd.Run(context.Background(), os.Args); err != nil {
 		fmt.Fprintln(appCmd.ErrWriter, "error: ", err)
 		os.Exit(1)
