+++ ctxcat/src/gitignore/ignore.go
