+++ ctxcat/src/gitignore/ignore_test.go
